+++ jgclark.Dashboard/src/react/components/ItemGrid.jsx
@@ -2,7 +2,7 @@
 //--------------------------------------------------------------------------
 // A grid layout for items within a section.
 // Called by ItemGrid component.
-// Last updated 2025-01-27 for v2.1.6
+// Last updated 2025-02-23 for v2.2.0.a3
 //--------------------------------------------------------------------------
 
 import React from 'react'
@@ -12,7 +12,7 @@ import ItemRow from './ItemRow.jsx'
 import { logDebug, logInfo } from '@helpers/react/reactDev.js'
 
 // Set to true to see some subtle shading of section backgrounds
-const showColoredBackgrounds = false
+// const showColoredBackgrounds = false
 
 type Props = {
   items: Array<TSectionItem>,
@@ -27,13 +27,15 @@ function ItemGrid({ items, thisSection }: Props): React$Node {
       ))
     : []
 
+  // Add a subtle green background colour to the section if there are no items, or if the first item is a congrats message,
+  // or if the section has asked for a coloured background.
   const sectionBackgroundColor =
     items.length === 0 || items[0].itemType === 'itemCongrats'
       ? `color-mix(in srgb, var(--bg-main-color), green 4%)`
-      : showColoredBackgrounds && thisSection.sectionTitleColorPart
-      ? `color-mix(in srgb, var(--bg-main-color), var(--fg-${thisSection.sectionTitleColorPart}) 4%)`
+      : (thisSection.showColoredBackground && thisSection.sectionTitleColorPart)
+        ? `color-mix(in srgb, var(--bg-main-color), var(--fg-${thisSection.sectionTitleColorPart}) 4%)`
       : 'var(--bg-main-color)'
-  // logDebug('ItemGrid', `sectionBackgroundColor: ${sectionBackgroundColor} from ${String(tasksToShow.length)} items`)
+  if (sectionBackgroundColor !== 'var(--bg-main-color)') logInfo('ItemGrid', `sectionBackgroundColor: ${sectionBackgroundColor} from ${String(items.length)} items`)
 
   return (
     <div className="sectionItemsGrid" id={`${thisSection.ID}-Section`} style={{ backgroundColor: sectionBackgroundColor }}>